refactor: 사용하지 않은 폴더 삭제
