+++ src/components/SimpleTravelProductCard.tsx
